Extract printAssignment
